BACKPORT: FROMLIST: pwm: Add a software PWM which toggles a GPIO from a high-resolution timer
Add a software PWM which toggles a GPIO from a high-resolution timer.

This will naturally not be as accurate or as efficient as a hardware
PWM, but it is useful in some cases.  I have for example used it for
evaluating LED brightness handling (via leds-pwm) on a board where the
LED was just hooked up to a GPIO, and for a simple verification of the
timer frequency on another platform.

Since high-resolution timers are used, sleeping gpio chips are not
supported and are rejected in the probe function.

Change-Id: Idbd041a5ab40530886192a913a092643646896ad
Signed-off-by: Vincent Whitchurch <vincent.whitchurch@axis.com>
Co-developed-by: Stefan Wahren <wahrenst@gmx.net>
Signed-off-by: Stefan Wahren <wahrenst@gmx.net>
Signed-off-by: Damon Ding <damon.ding@rock-chips.com>
(cherry picked from https://patchwork.ozlabs.org/project/linux-gpio/patch/20240204220851.4783-3-wahrenst@gmx.net/)