media: i2c: otp_eeprom: remove pd_offset in pdaf part
When the command rkisp_3A_server from the package camera-engine-rkisp[0][1]
is running, it will execute an ioctl with cmd RKMODULE_GET_MODULE_INFO and
a rkmodule_pdaf_inf parameter. Because the package is compiled with the
struct rkmodule_pdaf_inf without pd_offset, the ioctl will fail and
rkisp_3A_server will fail to start.

The structure of rkmodule_pdaf_inf is modified, and some driver code is
modified to ensure the code can be compiled without error.

The change is based on the commit of 31a0da2dca715c65a81fd86114cd14ef5ae84657

[0]: url: https://radxa-repo.github.io/op1-bookworm-test/pool/main/c/camera-engine-rkisp/camera_engine_rkisp_v2.3.0_arm64.deb
[1]: sha256: 2f921c70b310b81fc1af195ce95650ec0f02962de969de53cf80e6aab79e1271

Signed-off-by: Zhaoming Luo <luozhaoming@radxa.com>